refactor and fix types/bugs
